grouping

+++ lib/models/cart.dart
@@ -3,7 +3,7 @@ class CartItem {
   final String name;
   final double price;
   int qty;
-  double discountApplied;
+  int discountApplied;
 
   CartItem({required this.id, required this.name, required this.price, this.qty = 1, this.discountApplied = 0});
